test: move Favoris test to FicheJeune.page.test.tsx
